Minor bug fixed (segfault)
The older code sometimes gave a segmentation fault due to a voronoi
vertex being present outside the image boundaries. This caused the
program to access a point outside the image frame when checking
for collision. Now we check if the point is within the image frame
and then check if it colliding.