Bug fix in VoronoiPlanner
After filtering the voronoi graph, the planner was unable to find
the path in graph because it found the vertex nearest to start / goal
points by looking at the vertices in voronoi diagram and not the
voronoi graph. But these vertices had no edges connecting them to rest
of the graph as they were discarded while filtering.