Bugfix in potential_field_interpolator
Potential field interpolator got stuck in an infinite loop when the
result from 2 calls of max_descent_direction resulted in a net zero
movement. This caused the potential_field_interpolator to fail in
several cases. This has been fixed by remembering the old movement
direction and stopping the interpolation whenever a local minima is
encountered.