Improvement in probabilistic graph builder
Probabilistic graph builder added points to make the graph connected
whenever it was not. It would ignore the users prescription of the
number of vertices in the graph. It now informs the user if graph is not
connected but does not add points on its own. However if the user does
not specify the number of points, it adds points till the graph is not
connected.