Bugfix in Planner class
The planner inserts terminal points in graph before performing a graph
search but in some cases when the graph wasn't sufficiently dense, it
would fails to insert terminal points. This would cause the application
to crash when graph search was performed because terminal points were
never inserted in the graph. This has now been fixed, the planner prints
a warning message whenever it cannot insert the points and returns an
empty path.